move extractTracks to utils file
